Refactor QRC20 * Replace qrc20.rs to coins directory from utxo * Add qrc20/qrc20_tests.rs * Add qrc20/rpc_electrum.rs
