+++ mm2src/coins/qrc20.rs
@@ -1,22 +1,42 @@
-use super::*;
 use crate::eth::{u256_to_big_decimal, wei_from_big_decimal, ERC20_CONTRACT, PAYMENT_STATE_SENT,
                  PAYMENT_STATE_UNINITIALIZED, SWAP_CONTRACT};
-use crate::utxo::rpc_clients::UtxoRpcClientOps;
-use crate::{SwapOps, ValidateAddressResult};
-use bitcrypto::sha256;
+use crate::qrc20::rpc_electrum::{ContractCallResult, Qrc20RpcOps, TxHistoryItem, TxReceipt};
+use crate::utxo::rpc_clients::{ElectrumClient, UnspentInfo, UtxoRpcClientEnum, UtxoRpcClientOps};
+use crate::utxo::utxo_common::{self, big_decimal_from_sat, HISTORY_TOO_LARGE_ERROR};
+use crate::utxo::{qtum, sign_tx, utxo_arc_from_conf_and_request, ActualTxFee, AdditionalTxData, FeePolicy,
+                  GenerateTransactionError, RequestTxHistoryResult, UtxoAddressFormat, UtxoArc, UtxoArcCommonOps,
+                  UtxoCoinCommonOps, UtxoTx, VerboseTransactionFrom, UTXO_LOCK};
+use crate::{FoundSwapTxSpend, HistorySyncState, MarketCoinOps, MmCoin, SwapOps, TradeFee, TransactionDetails,
+            TransactionEnum, TransactionFut, ValidateAddressResult, WithdrawFee, WithdrawRequest};
+use async_trait::async_trait;
+use bigdecimal::BigDecimal;
+use bitcrypto::{dhash160, sha256};
+use chain::TransactionOutput;
 use common::block_on;
 use common::executor::Timer;
-use common::jsonrpc_client::{JsonRpcClient, JsonRpcErrorType, JsonRpcRequest, RpcRes};
+use common::jsonrpc_client::{JsonRpcClient, JsonRpcError, JsonRpcErrorType, JsonRpcRequest, RpcRes};
+use common::mm_ctx::MmArc;
 use common::mm_metrics::MetricsArc;
 use ethabi::{Function, Token};
 use ethereum_types::{H160, U256};
+use futures::compat::Future01CompatExt;
 use futures::{FutureExt, TryFutureExt};
+use futures01::Future;
 use gstuff::now_ms;
-use rpc::v1::types::H160 as H160Json;
+use keys::bytes::Bytes as ScriptBytes;
+use keys::{Address as UtxoAddress, Public};
+use mocktopus::macros::*;
+use rpc::v1::types::{Bytes as BytesJson, Transaction as RpcTransaction, H160 as H160Json, H256 as H256Json};
+use script::{Builder as ScriptBuilder, Opcode, Script, TransactionInputSigner};
+use serde_json::{self as json, Value as Json};
 use serialization::deserialize;
+use serialization::serialize;
 use std::ops::Neg;
 use std::str::FromStr;
-use utxo_common::HISTORY_TOO_LARGE_ERROR;
+
+#[cfg(test)] mod qrc20_tests;
+pub mod rpc_electrum;
+mod swap_ops;
 
 const QRC20_GAS_LIMIT_DEFAULT: u64 = 100_000;
 const QRC20_GAS_PRICE_DEFAULT: u64 = 40;
@@ -29,148 +49,6 @@ const QRC20_APPROVE_TOPIC: &str = "8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b22
 /// Keccak-256 hash of "PaymentSent" event generated by swap contract.
 #[allow(dead_code)]
 const QRC20_PAYMENT_SENT: &str = "ccc9c05183599bd3135da606eaaf535daffe256e9de33c048014cffcccd4ad57";
-
-#[derive(Debug, Deserialize, Eq, PartialEq)]
-pub struct TokenInfo {
-    name: String,
-    decimals: u8,
-    total_supply: u64,
-    symbol: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ExecutionResult {
-    pub output: BytesJson,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ContractCallResult {
-    address: H160Json,
-    #[serde(rename = "executionResult")]
-    pub execution_result: ExecutionResult,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct TxHistoryItem {
-    pub tx_hash: H256Json,
-    pub height: i64,
-    pub log_index: i64,
-}
-
-/// The structure is the same as Qtum Core RPC gettransactionreceipt returned data.
-/// https://docs.qtum.site/en/Qtum-RPC-API/#gettransactionreceipt
-#[derive(Debug, Deserialize)]
-pub struct TxReceipt {
-    /// Hash of the block this transaction was included within.
-    #[serde(rename = "blockHash")]
-    block_hash: H256Json,
-    /// Number of the block this transaction was included within.
-    #[serde(rename = "blockNumber")]
-    block_number: i64,
-    /// Transaction hash.
-    #[serde(rename = "transactionHash")]
-    transaction_hash: H256Json,
-    /// Index within the block.
-    #[serde(rename = "transactionIndex")]
-    transaction_index: i64,
-    /// Index within the outputs.
-    #[serde(rename = "outputIndex")]
-    output_index: i64,
-    /// 20 bytes，the sender address of this tx.
-    from: String,
-    /// 20 bytes，the receiver address of this tx. if this  address is created by a contract, return null.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    to: Option<String>,
-    /// The total amount of gas used after execution of the current transaction.
-    #[serde(rename = "cumulativeGasUsed")]
-    cumulative_gas_used: u64,
-    /// The gas cost alone to execute the current transaction.
-    #[serde(rename = "gasUsed")]
-    gas_used: i64,
-    /// Contract address created, or `None` if not a deployment.
-    #[serde(rename = "contractAddress")]
-    contract_address: Option<String>,
-    /// Logs generated within this transaction.
-    log: Vec<LogEntry>,
-    /// Whether corresponding contract call (specified in UTXO outputs[output_index]) was failed.
-    /// If None or Some("None") - completed, else failed.
-    excepted: Option<String>,
-    #[serde(rename = "exceptedMessage")]
-    excepted_message: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct LogEntry {
-    /// Contract address.
-    address: String,
-    /// Vector of 0x-prefixed hex strings with length of 64.
-    topics: Vec<String>,
-    /// In other words the data means a transaction value.
-    data: String,
-}
-
-/// QTUM specific RPC ops
-pub trait QtumRpcOps {
-    /// This can be used to get the basic information(name, decimals, total_supply, symbol) of a QRC20 token.
-    /// https://github.com/qtumproject/qtum-electrumx-server/blob/master/docs/qrc20-integration.md#blockchaintokenget_infotoken_address
-    fn blockchain_token_get_info(&self, token_addr: &H160Json) -> RpcRes<TokenInfo>;
-
-    fn blockchain_contract_call(&self, contract_addr: &H160Json, data: BytesJson) -> RpcRes<ContractCallResult>;
-
-    /// this can be used to retrieve QRC20 token transfer history, params are the same as blockchain.contract.event.subscribe,
-    /// and it returns a list of map{tx_hash, height, log_index}, where log_index is the position for this event log in its transaction.
-    /// https://github.com/qtumproject/qtum-electrumx-server/blob/master/docs/qrc20-integration.md#blockchaincontracteventget_historyhash160-contract_addr-topic
-    fn blockchain_contract_event_get_history(
-        &self,
-        address: &H160Json,
-        contract_addr: &H160Json,
-        topic: &str,
-    ) -> RpcRes<Vec<TxHistoryItem>>;
-
-    fn blochchain_transaction_get_receipt(&self, hash: &H256Json) -> RpcRes<Vec<TxReceipt>>;
-}
-
-impl QtumRpcOps for ElectrumClient {
-    fn blockchain_token_get_info(&self, token_addr: &H160Json) -> RpcRes<TokenInfo> {
-        rpc_func!(self, "blockchain.token.get_info", token_addr)
-    }
-
-    fn blockchain_contract_call(&self, contract_addr: &H160Json, data: BytesJson) -> RpcRes<ContractCallResult> {
-        let sender = "";
-        rpc_func!(self, "blockchain.contract.call", contract_addr, data, sender)
-    }
-
-    fn blockchain_contract_event_get_history(
-        &self,
-        address: &H160Json,
-        contract_addr: &H160Json,
-        topic: &str,
-    ) -> RpcRes<Vec<TxHistoryItem>> {
-        rpc_func!(
-            self,
-            "blockchain.contract.event.get_history",
-            address,
-            contract_addr,
-            topic
-        )
-    }
-
-    fn blochchain_transaction_get_receipt(&self, hash: &H256Json) -> RpcRes<Vec<TxReceipt>> {
-        rpc_func!(self, "blochchain.transaction.get_receipt", hash)
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub struct Qrc20FeeDetails {
-    /// Coin name
-    coin: String,
-    /// Standard UTXO miner fee based on transaction size
-    miner_fee: BigDecimal,
-    /// in satoshi
-    gas_limit: u64,
-    gas_price: u64,
-    total_gas_fee: BigDecimal,
-}
 
 pub async fn qrc20_coin_from_conf_and_request(
     ctx: &MmArc,
@@ -222,7 +100,7 @@ enum RpcContractCallType {
 #[derive(Clone, Debug, PartialEq)]
 pub struct ContractCallOutput {
     pub value: u64,
-    pub script_pubkey: Bytes,
+    pub script_pubkey: ScriptBytes,
     pub gas_limit: u64,
     pub gas_price: u64,
 }
@@ -277,9 +155,9 @@ impl Qrc20Coin {
         Ok(try_s!(function.decode_output(&result.execution_result.output)))
     }
 
-    pub fn utxo_address_from_qrc20(&self, address: H160) -> Address {
+    pub fn utxo_address_from_qrc20(&self, address: H160) -> UtxoAddress {
         let utxo = self.as_ref();
-        Address {
+        UtxoAddress {
             prefix: utxo.pub_addr_prefix,
             t_addr_prefix: utxo.pub_t_addr_prefix,
             hash: address.0.into(),
@@ -287,7 +165,7 @@ impl Qrc20Coin {
         }
     }
 
-    pub fn utxo_address_from_raw_pubkey(&self, pubkey: &[u8]) -> Result<Address, String> {
+    pub fn utxo_address_from_raw_pubkey(&self, pubkey: &[u8]) -> Result<UtxoAddress, String> {
         Ok(try_s!(utxo_common::address_from_raw_pubkey(
             pubkey,
             self.utxo_arc.pub_addr_prefix,
@@ -551,7 +429,7 @@ impl Qrc20Coin {
                 },
                 e => ERRL!("{}", e),
             })?;
-        let prev_script = Builder::build_p2pkh(&self.utxo_arc.my_address.hash);
+        let prev_script = ScriptBuilder::build_p2pkh(&self.utxo_arc.my_address.hash);
         let signed = try_s!(sign_tx(
             unsigned,
             &self.utxo_arc.key_pair,
@@ -762,7 +640,7 @@ impl UtxoCoinCommonOps for Qrc20Coin {
 
     async fn get_htlc_spend_fee(&self) -> Result<u64, String> { utxo_common::get_htlc_spend_fee(self).await }
 
-    fn addresses_from_script(&self, script: &Script) -> Result<Vec<Address>, String> {
+    fn addresses_from_script(&self, script: &Script) -> Result<Vec<UtxoAddress>, String> {
         utxo_common::addresses_from_script(&self.utxo_arc, script)
     }
 
@@ -770,15 +648,15 @@ impl UtxoCoinCommonOps for Qrc20Coin {
 
     fn my_public_key(&self) -> &Public { self.utxo_arc.key_pair.public() }
 
-    fn display_address(&self, address: &Address) -> Result<String, String> {
+    fn display_address(&self, address: &UtxoAddress) -> Result<String, String> {
         utxo_common::display_address(&self.utxo_arc, address)
     }
 
-    fn try_address_from_str(&self, from: &str) -> Result<Address, String> {
+    fn try_address_from_str(&self, from: &str) -> Result<UtxoAddress, String> {
         utxo_common::try_address_from_str(&self.utxo_arc, from)
     }
 
-    fn address_from_str(&self, address: &str) -> Result<Address, String> {
+    fn address_from_str(&self, address: &str) -> Result<UtxoAddress, String> {
         utxo_common::address_from_str(&self.utxo_arc, address)
     }
 
@@ -809,7 +687,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
         &self,
         unsigned: TransactionInputSigner,
         data: AdditionalTxData,
-        my_script_pub: Bytes,
+        my_script_pub: ScriptBytes,
     ) -> Result<(TransactionInputSigner, AdditionalTxData), String> {
         utxo_common::calc_interest_if_required(self, unsigned, data, my_script_pub).await
     }
@@ -817,7 +695,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
     fn p2sh_spending_tx(
         &self,
         prev_transaction: UtxoTx,
-        redeem_script: Bytes,
+        redeem_script: ScriptBytes,
         outputs: Vec<TransactionOutput>,
         script_data: Script,
         sequence: u32,
@@ -834,7 +712,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
 
     fn ordered_mature_unspents(
         &self,
-        address: &Address,
+        address: &UtxoAddress,
     ) -> Box<dyn Future<Item = Vec<UnspentInfo>, Error = String> + Send> {
         Box::new(
             utxo_common::ordered_mature_unspents(self.clone(), address.clone())
@@ -1417,9 +1295,9 @@ pub fn qrc20_addr_from_str(address: &str) -> Result<H160, String> {
     Ok(try_s!(json::from_str(&format!("\"{}\"", address))))
 }
 
-pub fn qrc20_addr_from_utxo_addr(address: Address) -> H160 { address.hash.take().into() }
+pub fn qrc20_addr_from_utxo_addr(address: UtxoAddress) -> H160 { address.hash.take().into() }
 
-fn utxo_addr_into_rpc_format(address: Address) -> H160Json { address.hash.take().into() }
+fn utxo_addr_into_rpc_format(address: UtxoAddress) -> H160Json { address.hash.take().into() }
 
 fn qrc20_addr_into_rpc_format(address: &H160) -> H160Json { address.to_vec().as_slice().into() }
 
@@ -1651,8 +1529,20 @@ fn find_swap_contract_call_with_swap_id(
     None
 }
 
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct Qrc20FeeDetails {
+    /// Coin name
+    coin: String,
+    /// Standard UTXO miner fee based on transaction size
+    miner_fee: BigDecimal,
+    /// in satoshi
+    gas_limit: u64,
+    gas_price: u64,
+    total_gas_fee: BigDecimal,
+}
+
 async fn qrc20_withdraw(coin: Qrc20Coin, req: WithdrawRequest) -> Result<TransactionDetails, String> {
-    let to_addr = try_s!(Address::from_str(&req.to));
+    let to_addr = try_s!(UtxoAddress::from_str(&req.to));
     let is_p2pkh =
         to_addr.prefix == coin.utxo_arc.pub_addr_prefix && to_addr.t_addr_prefix == coin.utxo_arc.pub_t_addr_prefix;
     let is_p2sh = to_addr.prefix == coin.utxo_arc.p2sh_addr_prefix
@@ -2323,7 +2213,7 @@ fn generate_contract_call_script_pubkey(
     let gas_limit = encode_contract_number(gas_limit as i64);
     let gas_price = encode_contract_number(gas_price as i64);
 
-    Ok(Builder::default()
+    Ok(ScriptBuilder::default()
         .push_opcode(Opcode::OP_4)
         .push_bytes(&gas_limit)
         .push_bytes(&gas_price)
@@ -2406,151 +2296,4 @@ fn extract_token_addr_from_script(script: &Script) -> Result<H160, String> {
     }
 
     Ok(instruction.data.ok_or(ERRL!("An empty contract call data"))?.into())
-}
-
-#[cfg(test)]
-mod qtum_tests {
-    use super::*;
-    use crate::{eth::u256_to_big_decimal, utxo::utxo_tests::electrum_client_for_test};
-    use ethabi::Token;
-    use keys::Address;
-
-    #[test]
-    fn blockchain_token_get_info() {
-        let client = electrum_client_for_test(&["95.217.83.126:10001"]);
-        let addr = hex::decode("d362e096e873eb7907e205fadc6175c6fec7bc44").unwrap();
-        let expected_token_info = TokenInfo {
-            name: "TEST".into(),
-            decimals: 8,
-            total_supply: 1000000000000000,
-            symbol: "ARTEM".into(),
-        };
-
-        let actual_token_info = client
-            .blockchain_token_get_info(&addr.as_slice().into())
-            .wait()
-            .unwrap();
-        assert_eq!(expected_token_info, actual_token_info);
-    }
-
-    #[test]
-    fn get_token_balance_using_contract_call() {
-        let client = electrum_client_for_test(&["95.217.83.126:10001"]);
-        let token_addr = hex::decode("d362e096e873eb7907e205fadc6175c6fec7bc44").unwrap();
-        let our_addr: Address = "qKEDGuogDhtH9zBnc71QtqT1KDamaR1KJ3".parse().unwrap();
-        log!((our_addr.prefix));
-        let function = unwrap!(ERC20_CONTRACT.function("balanceOf"));
-        let data = unwrap!(function.encode_input(&[Token::Address(our_addr.hash.take().into()),]));
-        let balance = client
-            .blockchain_contract_call(&token_addr.as_slice().into(), data.into())
-            .wait()
-            .unwrap();
-        let tokens = function.decode_output(&balance.execution_result.output).unwrap();
-        let balance = match tokens[0] {
-            Token::Uint(bal) => bal,
-            _ => panic!("Expected Uint, got {:?}", tokens[0]),
-        };
-        let balance = u256_to_big_decimal(balance, 8).unwrap();
-        assert_eq!(balance, "139.00000".parse().unwrap());
-    }
-
-    #[test]
-    fn test_encode_decode_contract_number() {
-        let numbers = vec![
-            // left is source number, right is expected encoded array
-            (0i64, vec![]),
-            (1, vec![1]),
-            (-1, vec![129]),
-            (40, vec![40]),
-            (-40, vec![168]),
-            (-127, vec![255]),
-            (127, vec![127]),
-            (-128, vec![128, 128]),
-            (128, vec![128, 0]),
-            (255, vec![255, 0]),
-            (-255, vec![255, 128]),
-            (256, vec![0, 1]),
-            (-256, vec![0, 129]),
-            (2500000, vec![160, 37, 38]),
-            (-2500000, vec![160, 37, 166]),
-            (i64::max_value(), vec![255, 255, 255, 255, 255, 255, 255, 127]),
-            (i64::min_value(), vec![0, 0, 0, 0, 0, 0, 0, 128, 128]),
-            (Opcode::OP_4 as i64, vec![84]),
-            (Opcode::OP_CALL as i64, vec![194, 0]),
-        ];
-
-        for (source, encoded) in numbers {
-            println!("{}", source);
-            let actual_encoded = encode_contract_number(source);
-            assert_eq!(actual_encoded, encoded);
-            let actual_decoded = unwrap!(decode_contract_number(&encoded));
-            assert_eq!(actual_decoded, source);
-        }
-    }
-
-    #[test]
-    fn test_extract_gas_limit_gas_price() {
-        let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-
-        let expected_gas_limit = 2_500_000;
-        let actual = unwrap!(extract_gas_from_script(&script, ExtractGasEnum::GasLimit));
-        assert_eq!(actual, expected_gas_limit);
-
-        let expected_gas_price = 40;
-        let actual = unwrap!(extract_gas_from_script(&script, ExtractGasEnum::GasPrice));
-        assert_eq!(actual, expected_gas_price);
-    }
-
-    #[test]
-    fn test_extract_contract_call() {
-        let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-
-        let to_addr: Address = "qHmJ3KA6ZAjR9wGjpFASn4gtUSeFAqdZgs".into();
-        let to_addr = qrc20_addr_from_utxo_addr(to_addr);
-        let amount: U256 = 1000000000.into();
-        let function = ERC20_CONTRACT.function("transfer").unwrap();
-        let expected = function
-            .encode_input(&[Token::Address(to_addr), Token::Uint(amount)])
-            .unwrap();
-
-        let actual = unwrap!(extract_contract_call_from_script(&script));
-        assert_eq!(actual, expected);
-
-        // TX b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0
-        let tx = unwrap!(hex::decode("01000000029ba0865fc62aac1f5f1a4aac3c9f54ff3d74211030bf6eb41e870b30297bd3fc010000006a47304402201808cbc98036ea63d32e858f776c722897d3f4b670744594deba25b69128d0ba02207b3f86f0ab6b6fa0ff581dc7be33af034c6004f3537a2f96c4ddf3ed0130defc012103693bff1b39e8b5a306810023c29b95397eb395530b106b1820ea235fd81d9ce9ffffffff63574ffa1e8edd8af8b08f3c1d8e5f33170772c38631a50ddc29c16d74c762f6020000006b483045022100db6cf963f6be56f7c6004ede74d452b2c5932eb6b12094fe67fa4ff0b6f4406e02207acff9163588a0c58fa009f5e876f90817b4006c303095d9b6425aae2922a485012103693bff1b39e8b5a306810023c29b95397eb395530b106b1820ea235fd81d9ce9ffffffff040000000000000000625403a08601012844095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a000000000000000000000000000000000000000000000000000000000000000014d362e096e873eb7907e205fadc6175c6fec7bc44c20000000000000000625403a08601012844095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000001312d0014d362e096e873eb7907e205fadc6175c6fec7bc44c20000000000000000e35403a0860101284cc49b415b2a65e285b98480fd7de696e9fb5bcb68ec9468dd906c683e38cabb8f39905675fa0000000000000000000000000000000000000000000000000000000001312d00000000000000000000000000d362e096e873eb7907e205fadc6175c6fec7bc440000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde30101010101010101010101010101010101010101000000000000000000000000000000000000000000000000000000000000000000000000000000005f6d80b814ba8b71f3544b93e2f681f996da519a98ace0107ac2e52fdd05000000001976a9149e032d4b0090a11dc40fe6c47601499a35d55fbb88ac82816d5f"));
-        let utxo_tx: UtxoTx = unwrap!(deserialize(tx.as_slice()));
-
-        // first output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-        // `approve` to 0 contract call
-        let expected = unwrap!(hex::decode("095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000000000000"));
-        let script = utxo_tx.outputs[0].script_pubkey.clone().into();
-
-        let actual = unwrap!(extract_contract_call_from_script(&script));
-        assert_eq!(actual, expected);
-
-        // second output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-        // `approve` to 20000000 contract call
-        let expected = unwrap!(hex::decode("095ea7b3000000000000000000000000ba8b71f3544b93e2f681f996da519a98ace0107a0000000000000000000000000000000000000000000000000000000001312d00"));
-        let script = utxo_tx.outputs[1].script_pubkey.clone().into();
-
-        let actual = unwrap!(extract_contract_call_from_script(&script));
-        assert_eq!(actual, expected);
-
-        // third output in "b11a262380657310abf01f8abe117da2c2adf788ab1fa0fa29da4ab505fc00c0"
-        // `erc20Payment` 20000000 amount contract call
-        let expected = unwrap!(hex::decode("9b415b2a65e285b98480fd7de696e9fb5bcb68ec9468dd906c683e38cabb8f39905675fa0000000000000000000000000000000000000000000000000000000001312d00000000000000000000000000d362e096e873eb7907e205fadc6175c6fec7bc440000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde30101010101010101010101010101010101010101000000000000000000000000000000000000000000000000000000000000000000000000000000005f6d80b8"));
-        let script = utxo_tx.outputs[2].script_pubkey.clone().into();
-
-        let actual = unwrap!(extract_contract_call_from_script(&script));
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn extract_token_addr() {
-        let script: Script = "5403a02526012844a9059cbb0000000000000000000000000240b898276ad2cc0d2fe6f527e8e31104e7fde3000000000000000000000000000000000000000000000000000000003b9aca0014d362e096e873eb7907e205fadc6175c6fec7bc44c2".into();
-        let expected = qrc20_addr_from_str("0xd362e096e873eb7907e205fadc6175c6fec7bc44").unwrap();
-
-        let actual = unwrap!(extract_token_addr_from_script(&script));
-        assert_eq!(actual, expected);
-    }
 }
